+++ CHANGELOG.md
@@ -6,3 +6,4 @@
 - Introduce scan mode: `--double-sided` => `--mode=double`, `--no-adf` => `--mode=multi`, new mode `--mode=one`
 - Respect the deleted image during the preview
 - Add crop status image
+- Add min black to determine if a content box is valid